Add option for client endpoints to use annotated classes
Add an option to configure a client endpoint using an annotated class in
addition to the existing options of using a pre-created annotated
instance, an endpoint class or endpoint instance.
The naming convention for adding the client endpoint was inconsistent so
this commit includes refactoring the configuration of client endpoints
to make the naming convention consistent.